refactor: Consolidated all factories into core.factories [#50]
+ Used import path string to prevent circular dependencies
