--- django_api/django_api/apps/account/factories.py
@@ -1,44 +1,0 @@
-from django.contrib.auth.models import Group
-
-import factory
-from factory import fuzzy
-
-from account.models import User, UserProfile
-
-
-class UserFactory(factory.django.DjangoModelFactory):
-    username = factory.Sequence(lambda n: "user_%d" % n)
-    email = factory.Sequence(lambda n: "user{}@notanemail.com".format(n))
-    password = factory.PostGenerationMethodCall('set_password', 'test')
-
-    profile = factory.RelatedFactory(ProfileFactory, 'user')
-
-    @classmethod
-    def _generate(cls, create, attrs):
-        """Override the default _generate() to disable the post-save signal."""
-
-        # Note: If the signal was defined with a dispatch_uid, include that in both calls.
-        post_save.disconnect(UserProfile.create_user_profile, User)
-        user = super(UserFactory, cls)._generate(create, attrs)
-        post_save.connect(UserProfile.create_user_profile, User)
-        return user
-
-    @factory.post_generation
-    def groups(self, create, extracted, **kwargs):
-        group, created = Group.objects.get_or_create(name='UNICEF User')
-        self.groups.add(group)
-
-    class Meta:
-        model = User
-
-
-class UserProfileFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = UserProfile
-
-
-class GroupFactory(factory.django.DjangoModelFactory):
-    name = "UNICEF User"
-
-    class Meta:
-        model = Group
