+++ django_api/requirements/base.txt
@@ -13,3 +13,4 @@ django-mptt==0.8.6
 astroid==1.5.2
 isort==4.2.5
 pylint==1.7.1
+factory_boy==2.8.1
